// exercices/1.3/basic-ts-api-boilerplate-main/types.ts
interface Pizza {
  id: number;
  title: string;
  content: string;
}

interface PizzaToUpdate {
  title?: string;
  content?: string;
}

interface Film {
  id: number;
  title: string;
  director: string;
  duration: number;
  budget?: number;
  description?: string;
  imageUrl?: string;
}


type NewPizza = Omit<Pizza, "id">;
type NewFilm = Omit<Film, "id">;

export type { Pizza, NewPizza, PizzaToUpdate, Film, NewFilm };